# src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: 뭉치장 API
  description: |
    유행 음식 공동구매 플랫폼 뭉치장의 API 명세입니다.

    ## 인증
    로그인이 필요한 API는 `Authorization: Bearer {accessToken}` 헤더를 포함해야 합니다.

    ## 공통 응답 형식
    모든 응답은 아래 구조를 따릅니다.
    ```json
    { "success": true, "data": { ... }, "error": null }
    { "success": false, "data": null, "error": { "code": "ERROR_CODE", "message": "설명", "detail": "상세" } }
    ```
  version: v1
servers:
  - url: https://api.moongchijang.com
    description: 운영 서버
  - url: https://api.moongchijang.com/dev
    description: 개발 서버
  - url: http://localhost:8080
    description: 로컬 서버
...
paths:
...
  /api/v1/group-buys/{groupBuyId}:
    get:
      tags: [GroupBuy]
      summary: 공구 상세 조회
      description: 공구의 상품·매장·픽업·이미지 정보를 조회한다. 비로그인 허용.
      parameters:
        - $ref: '#/components/parameters/GroupBuyId'
      responses:
        '200':
          description: 공구 상세
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseGroupBuyDetailResponse'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/v1/group-buy-requests:
    post:
      tags: [GroupBuyRequest]
      summary: 공구 개설 요청 제출
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GroupBuyRequestCreate'
      responses:
        '201':
          description: 요청 접수 완료
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseRequestId'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
    get:
      tags: [GroupBuyRequest]
      summary: 내 공구 요청 목록 조회
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 내 요청 목록
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseGroupBuyRequestList'
  /api/v1/group-buy-requests/{requestId}:
    get:
      tags: [GroupBuyRequest]
      summary: 공구 요청 상세 조회
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/RequestId'
      responses:
        '200':
          description: 요청 상세
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseGroupBuyRequestDetail'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/v1/group-buys/{groupBuyId}/wishlist:
    post:
      tags: [Wishlist]
      summary: 찜 추가
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/GroupBuyId'
      responses:
        '201':
          $ref: '#/components/responses/SuccessNoData'
        '401':
          $ref: '#/components/responses/Unauthorized'
    delete:
      tags: [Wishlist]
      summary: 찜 해제
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/GroupBuyId'
      responses:
        '200':
          $ref: '#/components/responses/SuccessNoData'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/wishlists:
    get:
      tags: [Wishlist]
      summary: 찜 목록 조회
      security:
        - bearerAuth: []
      parameters:
        - name: filter
          in: query
          schema:
            type: string
            enum: [ALL, CLOSING_SOON, OPEN]
            default: ALL
          description: ALL=전체 / CLOSING_SOON=마감임박(D-3) / OPEN=모집중
        - name: excludeClosed
          in: query
          schema:
            type: boolean
            default: false
          description: true면 마감 공고 제외, false면 마감 포함
        - name: sort
          in: query
          schema:
            type: string
            enum: [LATEST, DEADLINE]
            default: LATEST
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: 찜 목록
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseWishlistPage'

  # ─────────────────────────────────────────────────────────────
  # Participation & Payment
  # ─────────────────────────────────────────────────────────────
...
  /api/v1/owner/group-buy-requests/{requestId}:
    get:
      tags: [Owner]
      summary: 사장님 공구 개설 요청 상세 조회
      description: 사장님이 본인 매장 기준으로 제출한 공구 개설 요청 상세와 승인/반려 상태를 조회한다.
      security:
        - bearerAuth: []
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 사장님 공구 개설 요청 상세
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseOwnerGroupBuyRequestDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/v1/admin/group-buy-requests:
    get:
      tags: [Admin]
      summary: 공구 요청 목록 조회 (운영자)
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [ALL, IN_REVIEW, IN_CONTACT, OPENED, REJECTED]
            default: ALL
        - name: keyword
          in: query
          required: false
          schema:
            type: string
          description: 요청 ID, 상품명, 가게명, 요청자 닉네임/이메일/전화번호 검색어
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: 공구 요청 목록
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseAdminRequestPage'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/admin/group-buy-requests/{requestId}:
    get:
      tags: [Admin]
      summary: 공구 요청 상세 조회 (운영자)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/RequestId'
      responses:
        '200':
          description: 요청 상세
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseAdminRequestDetail'
  /api/v1/admin/group-buy-requests/{requestId}/status:
    patch:
      tags: [Admin]
      summary: 공구 요청 상태 변경
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/RequestId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminRequestStatusUpdate'
      responses:
        '200':
          description: 상태 변경 결과
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseGroupBuyRequestDetail'
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /api/v1/admin/settlements/dashboard:
    get:
      tags: [Admin]
      summary: 정산 현황 대시보드 조회 (운영자)
      description: 선택한 연월의 정산 완료 금액, 정산 예정 금액, 서비스 수수료, 총 거래액을 조회한다. 현재 서비스 수수료 정책은 0원이다.
      security:
        - bearerAuth: []
      parameters:
        - name: year
          in: query
          required: true
          schema:
            type: integer
            example: 2026
        - name: month
          in: query
          required: true
          schema:
            type: integer
            minimum: 1
            maximum: 12
            example: 5
      responses:
        '200':
          description: 정산 대시보드
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseAdminSettlementDashboard'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/admin/settlements:
    get:
      tags: [Admin]
      summary: 정산 현황 목록 조회 (운영자)
      description: 선택한 연월의 정산 예정, 정산 완료, 전체 목록을 조회한다. 현재 settlementId는 groupBuyId와 동일하다.
      security:
        - bearerAuth: []
      parameters:
        - name: year
          in: query
          required: true
          schema:
            type: integer
            example: 2026
        - name: month
          in: query
          required: true
          schema:
            type: integer
            minimum: 1
            maximum: 12
            example: 5
        - name: status
          in: query
          schema:
            type: string
            enum: [SCHEDULED, COMPLETED, ALL]
            default: ALL
          description: SCHEDULED=정산 예정, COMPLETED=정산 완료, ALL=전체
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
      responses:
        '200':
          description: 정산 현황 목록
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseAdminSettlementPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/admin/settlements/{settlementId}:
    get:
      tags: [Admin]
      summary: 정산 현황 상세 조회 (운영자)
      description: 목록에서 선택한 정산 건의 상세 팝업용 정보를 조회한다. 현재 settlementId는 groupBuyId와 동일하다.
      security:
        - bearerAuth: []
      parameters:
        - name: settlementId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 정산 상세
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseAdminSettlementDetail'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ─────────────────────────────────────────────────────────────
  # ImageUpload
  # ─────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    GroupBuyId:
      name: groupBuyId
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
    RequestId:
      name: requestId
      in: path
      required: true
      schema:
        type: integer
        format: int64
    Page:
      name: page
      in: query
      schema:
        type: integer
        default: 0
    Size:
      name: size
      in: query
      schema:
        type: integer
        default: 20
        maximum: 100
  responses:
    SuccessNoData:
      description: 성공 (데이터 없음)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponseEmpty'
    BadRequest:
      description: 잘못된 요청
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponseError'
    Unauthorized:
      description: 인증 실패
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponseError'
    Forbidden:
      description: 권한 없음
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponseError'
    NotFound:
      description: 리소스 없음
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponseError'
...
  schemas:
...
    ApiResponseGroupBuyDetailResponse:
      type: object
      required: [success, data, error]
      properties:
        success: { type: boolean, example: true }
        data:
          type: object
          required: [id, storeName, regionType, regionLabel, districtType, districtLabel, productName, productDescription, imageUrls, price, achievementRate, currentQuantity, targetQuantity, deadline, pickupDate, pickupTimeStart, pickupTimeEnd, pickupDateLabel, pickupDateTimeLabel, deadlineDateTimeLabel, pickupLocation, dDay, dDayLabel, isWishlisted, isClosed, isParticipated, canParticipate]
          properties:
            id: { type: integer, format: int64 }
            storeName: { type: string }
            regionType: { $ref: '#/components/schemas/RegionType' }
            regionLabel: { type: string, description: 시/도 한글 라벨, example: 서울 }
            districtType: { $ref: '#/components/schemas/DistrictType' }
            districtLabel: { type: string, description: 세부지역 한글 라벨, example: "강남 | 역삼 | 삼성" }
            productName: { type: string }
            productDescription: { type: string }
            thumbnailUrl: { type: string, nullable: true, description: 대표 썸네일 이미지 URL }
            imageUrls:
              type: array
              items: { type: string }
            price: { type: integer }
            achievementRate: { type: integer, description: 달성률 % (0~100) }
            currentQuantity: { type: integer }
            targetQuantity: { type: integer }
            maxQuantity: { type: integer, nullable: true }
            deadline: { type: string, format: date-time }
            pickupDate: { type: string, format: date }
            pickupTimeStart: { type: string, format: time }
            pickupTimeEnd: { type: string, format: time }
            pickupDateLabel: { type: string, description: 픽업 날짜 표시 문자열, example: "4/15(화)" }
            pickupDateTimeLabel: { type: string, description: 픽업일시 표시용 문자열 }
            deadlineDateTimeLabel: { type: string, description: 마감일시 표시용 문자열 }
            pickupLocation: { type: string }
            pickupLatitude: { type: number, format: double, nullable: true }
            pickupLongitude: { type: number, format: double, nullable: true }
            dDay: { type: integer }
            dDayLabel: { type: string, description: "마감 뱃지 문자열 (예: D-3)" }
            isWishlisted: { type: boolean }
            isClosed: { type: boolean }
            isParticipated: { type: boolean }
            canParticipate: { type: boolean }
        error: { nullable: true, example: null }
...
    GroupBuyRequestCreate:
      type: object
      required: [storeName, productName, desiredQuantity, desiredPickupDate]
      properties:
        storeName:
          type: string
          maxLength: 100
        storeAddress:
          type: string
          maxLength: 200
          nullable: true
          description: 매장 주소. roadAddress가 함께 전달되면 서버는 roadAddress를 우선 저장한다.
        placeId:
          type: string
          maxLength: 100
          nullable: true
          description: 네이버 장소 고유 ID
        roadAddress:
          type: string
          maxLength: 200
          nullable: true
          description: 네이버 장소 도로명 주소
        lotAddress:
          type: string
          maxLength: 200
          nullable: true
          description: 네이버 장소 지번 주소
        latitude:
          type: number
          format: double
          nullable: true
          description: 네이버 장소 위도
        longitude:
          type: number
          format: double
          nullable: true
          description: 네이버 장소 경도
        productName:
          type: string
          maxLength: 100
        desiredQuantity:
          type: integer
          minimum: 1
        desiredPickupDate:
          type: string
          format: date
          description: 오늘 이후 날짜
        additionalNote:
          type: string
          maxLength: 500
          nullable: true
    ApiResponseRequestId:
      type: object
      required: [success, data, error]
      properties:
        success: { type: boolean, example: true }
        data:
          type: object
          required: [requestId]
          properties:
            requestId: { type: integer, format: int64 }
        error: { nullable: true, example: null }
...
    ApiResponseGroupBuyRequestList:
      type: object
      required: [success, data, error]
      properties:
        success: { type: boolean, example: true }
        data:
          type: array
          items:
            $ref: '#/components/schemas/GroupBuyRequestDetail'
        error: { nullable: true, example: null }
    ApiResponseGroupBuyRequestDetail:
      type: object
      required: [success, data, error]
      properties:
        success: { type: boolean, example: true }
        data:
          $ref: '#/components/schemas/GroupBuyRequestDetail'
        error: { nullable: true, example: null }

    # ── Wishlist ───────────────────────────────────────────────
    ApiResponseWishlistPage:
      type: object
      required: [success, data, error]
      properties:
        success: { type: boolean, example: true }
        data:
          type: object
          required: [content, totalElements, totalPages, number, size, urgentCount]
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/WishlistItemResponse'
            totalElements: { type: integer }
            totalPages: { type: integer }
            number: { type: integer }
            size: { type: integer }
            urgentCount:
              type: integer
              description: 마감 24시간 이내 건수 (0이면 배너 미노출)
        error: { nullable: true, example: null }

    # ── Participation ──────────────────────────────────────────
...
    ApiResponseOwnerGroupBuyRequestDetail:
      type: object
      required: [success, data, error]
      properties:
        success: { type: boolean, example: true }
        data:
          type: object
          required:
            - requestId
            - storeId
            - storeName
            - productName
            - productDescription
            - price
            - targetQuantity
            - maxQuantity
            - thumbnailUrl
            - imageUrls
            - deadline
            - pickupDate
            - pickupTimeStart
            - pickupTimeEnd
            - pickupLocation
            - status
          properties:
            requestId: { type: integer, format: int64, example: 101 }
            storeId: { type: integer, format: int64, example: 1 }
            storeName: { type: string, example: 뭉치장 베이커리 }
            productName: { type: string, example: 두쫀쿠 세트 }
            productDescription: { type: string, example: 소금빵 포함 }
            originalPrice: { type: integer, nullable: true, example: 12000 }
            price: { type: integer, example: 9900 }
            targetQuantity: { type: integer, example: 20 }
            maxQuantity: { type: integer, example: 50 }
            perUserLimit: { type: integer, nullable: true, example: 2 }
            thumbnailUrl: { type: string, example: "https://cdn.example.com/owner-requests/1.jpg" }
            imageUrls:
              type: array
              items:
                type: string
              example:
                - https://cdn.example.com/owner-requests/1.jpg
                - https://cdn.example.com/owner-requests/2.jpg
            deadline: { type: string, format: date-time, example: "2026-06-02T23:59:00" }
            pickupDate: { type: string, format: date, example: "2026-06-03" }
            pickupTimeStart: { type: string, format: time, example: "12:00:00" }
            pickupTimeEnd: { type: string, format: time, example: "18:00:00" }
            pickupLocation: { type: string, example: 서울 성동구 성수이로 1 }
            pickupContact: { type: string, nullable: true, example: "01012345678" }
            status:
              type: string
              enum: [PENDING, APPROVED, REJECTED]
            rejectionReason: { type: string, nullable: true, example: 매장 사정으로 어렵습니다 }
            approvedGroupBuyId: { type: integer, format: int64, nullable: true, example: 300 }
            reviewedAt: { type: string, format: date-time, nullable: true }
            requestedAt: { type: string, format: date-time, nullable: true }
        error: { nullable: true, example: null }
...
    ApiResponseAdminRequestPage:
      type: object
      required: [success, data, error]
      properties:
        success: { type: boolean, example: true }
        data:
          type: object
          required: [content, totalElements, totalPages, number, size]
          properties:
            content:
              type: array
              items:
                type: object
                required: [requestId, storeName, productName, desiredQuantity, desiredPickupDate, status, requesterId, actionable, createdAt]
                properties:
                  requestId: { type: integer, format: int64 }
                  storeName: { type: string }
                  productName: { type: string }
                  desiredQuantity: { type: integer }
                  desiredPickupDate: { type: string, format: date }
                  status:
                    type: string
                    enum: [IN_REVIEW, IN_CONTACT, OPENED, REJECTED]
                  requesterId: { type: integer, format: int64 }
                  requesterName: { type: string, nullable: true }
                  originalPrice:
                    type: integer
                    nullable: true
                    description: 승인 후 생성된 공구 정가. 승인 전 요청은 null
                  price:
                    type: integer
                    nullable: true
                    description: 승인 후 생성된 공구가. 승인 전 요청은 null
                  reviewElapsedMinutes:
                    type: integer
                    format: int64
                    nullable: true
                    description: 요청 생성 후 경과한 검토 시간(분)
                  actionable:
                    type: boolean
                    description: 운영자가 승인/반려 등 후속 작업을 할 수 있는 상태 여부
                  createdAt: { type: string, format: date-time, nullable: true }
            totalElements: { type: integer, format: int64 }
            totalPages: { type: integer }
            number: { type: integer }
            size: { type: integer }
        error: { nullable: true, example: null }
...